+++ resources/js/components/gallery/albumModule/thumbs/PhotoThumb.vue
@@ -87,7 +87,7 @@
 			<ThumbBadge
 				v-if="(albumsStore.rootRights?.can_highlight || albumStore.rights?.can_edit) && props.photo.is_highlighted"
 				class="bg-yellow-500"
-				icon="flag-fill"
+				pi="flag-fill"
 			/>
 			<ThumbBadge v-if="userStore.isLoggedIn && props.isCoverId" class="bg-yellow-500" icon="folder-cover" />
 			<ThumbBadge v-if="userStore.isLoggedIn && props.isHeaderId" class="bg-slate-400 hidden sm:block" pi="image" />
